Refactor for domain, business and API

+++ Source/Backend/ServiceMonitor/ServiceMonitor.Core/Business/Contracts/IAdministrationService.cs
@@ -1,5 +1,4 @@
 ﻿using System.Threading.Tasks;
-using ServiceMonitor.Core.Business.Responses;
 using ServiceMonitor.Core.Business.Responses.Contracts;
 using ServiceMonitor.Core.Domain;
 
@@ -7,6 +6,6 @@ namespace ServiceMonitor.Core.Business.Contracts
 {
     public interface IAdministrationService : IService
     {
-        Task<ISingleResponse<ServiceEnvironmentStatusLog>> CreateServiceEnvironmentStatusLogAsync(ServiceEnvironmentStatusLog entity, int? serviceEnvironmentID);
+        Task<ISingleResponse<ServiceEnvironmentStatusLog>> CreateServiceEnvironmentStatusLogAsync(ServiceEnvironmentStatusLog entity, short? serviceEnvironmentID);
     }
 }
