+++ Source/Backend/ServiceMonitor/ServiceMonitor.Core/Domain/Service.cs
@@ -6,14 +6,14 @@
         {
         }
 
-        public Service(int? id)
+        public Service(short? id)
         {
             ID = id;
         }
 
-        public int? ID { get; set; }
+        public short? ID { get; set; }
 
-        public int? ServiceCategoryID { get; set; }
+        public short? ServiceCategoryID { get; set; }
 
         public string Name { get; set; }
     }
